Machine Learning/Codes/Deep Learning/testLLM/requirements.txt

langchain_openai
langchain_core
langchain_community
langchain_text_splitters
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse_starlette
ipykernel
bs4
pypdf
chromadb
faiss-cpu